Get the axdebug extension building on amd64.  With lots of help from Steve Yin via [1727357] new VC8.0 build patch for AMD64 branch (inc axdebug)

+++ com/win32comext/axdebug/src/PyIDebugProperties.cpp
@@ -61,7 +61,7 @@ BOOL PyObject_AsDebugPropertyInfo(PyObject *ob, DebugPropertyInfo *p)
 		PyErr_SetString(PyExc_TypeError, "DebugProperty object must be a sequence");
 		return FALSE;
 	}
-	int len = PySequence_Length(ob);
+	Py_ssize_t len = PySequence_Length(ob);
 	if (len<1) {
 		PyErr_SetString(PyExc_ValueError, "DebugProperty sequence must be at least 1");
 		return FALSE;
